add new page finders base class for a few api endpoints to use.
