+++ lib/tankard/api/yeasts.rb
@@ -1,5 +1,5 @@
 require 'hashie'
-require 'tankard/api/utils/page_finders'
+require 'tankard/api/base/page_finders'
 
 module Tankard
   module Api
@@ -7,20 +7,15 @@ module Tankard
     #
     # @see http://www.brewerydb.com/developers/docs-endpoint/yeast_index
     # @author Matthew Shafer
-    class Yeasts
-      include Tankard::Api::Utils::PageFinders
+    class Yeasts < Tankard::Api::Base::PageFinders
       # @!parse include ::Enumerable
 
-      # Initializes a new object
+      # @!method initialize(request, options = {})
+      #   Initializes a new object
       #
-      # @param request [Tankard::Request]
-      # @param options [Hash]
-      # @return [Tankard::Api::Yeasts]
-      def initialize(request, options = {})
-        @http_client = request
-        @http_request_parameters = Hashie::Mash.new(options)
-        @http_request_uri = 'yeasts'
-      end
+      #   @param request [Tankard::Request]
+      #   @param options [Hash]
+      #   @return [Tankard::Api::Yeasts]
 
       # @!method each(&block)
       #   Calls the given block once for each yeast
@@ -31,20 +26,23 @@ module Tankard
       #   @raise [Tankard::Error::HttpError] when a status other than 200 or 401 is returned
       #   @raise [Tankard::Error::LoadError] when multi json is unable to decode json
 
-      # Specific page to request
+      # @!method page(number)
+      #   Specific page to request
       #
-      # @param number [Integer]
-      # @return [self] returns itself
-      def page(number)
-        @http_request_parameters[:p] = number
-        self
-      end
+      #   @param number [Integer]
+      #   @return [self] returns itself
+
+      # @!method params(options = {})
+      #   Additional parameters to send with the request
+      #
+      #   @param options [Hash]
+      #   @return [self] returns itself
 
     private
 
-      attr_reader :http_client
-      attr_reader :http_request_parameters
-      attr_reader :http_request_uri
+      def http_request_uri
+        'yeasts'
+      end
 
     end
   end
